# [SPARK-38505][SQL] Make partial aggregation adaptive