--- a/sql/core/src/main/java/org/apache/spark/sql/execution/UnsafeFixedWidthAggregationMap.java
+++ b/sql/core/src/main/java/org/apache/spark/sql/execution/UnsafeFixedWidthAggregationMap.java
@@ -227,6 +227,10 @@ public final class UnsafeFixedWidthAggregationMap {
     return map.getAvgHashProbesPerKey();
   }
 
+  public int getNumKeys() {
+    return map.numKeys();
+  }
+
   /**
    * Sorts the map's records in place, spill them to disk, and returns an [[UnsafeKVExternalSorter]]
    *