--- a/sql/core/src/main/scala/org/apache/spark/sql/execution/aggregate/HashMapGenerator.scala
+++ b/sql/core/src/main/scala/org/apache/spark/sql/execution/aggregate/HashMapGenerator.scala
@@ -75,6 +75,8 @@ abstract class HashMapGenerator(
        |
        |${generateRowIterator()}
        |
+       |${generateGetNumKeys()}
+       |
        |${generateClose()}
        |}
      """.stripMargin
@@ -132,6 +134,14 @@ abstract class HashMapGenerator(
     """
        |public void close() {
        |  batch.close();
+       |}
+     """.stripMargin
+  }
+
+  protected final def generateGetNumKeys(): String = {
+    s"""
+       |public int getNumKeys() {
+       |  return batch.numRows();
        |}
      """.stripMargin
   }