### What changes were proposed in this pull request?

This PR makes `HashAggregateExec` adaptively skip partial aggregation to avoid spilling if partial aggregation do not reduce the number of output rows too much. 

By setting `spark.sql.aggregate.adaptivePartialAggregationThreshold` to 0 this feature can be disabled.

### Why are the changes needed?

Improve partial aggregation phase performance and we can implement these 2 features after this PR:
1. SPARK-36245: Partial deduplicate the right side of left semi/anti join
2. SPARK-38506: Push partial aggregation through join

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Unit test and TPC-H 5T benchmark test.

SQL | Before this PR(Seconds) | After this PR(Seconds)
-- | -- | --
q15 | 66  | 51
q17 | 86 | 80
q18 | 129 | 122