@shenjiayu17 2022-12-09
Hi @wangyum. I'm very interested in this optimization on partial aggregation. But why does it need these child node limit? Do they make some influence on function or performance?
```
private[sql] lazy val isAdaptivePartialAggregationEnabled = {
    requiredChildDistributionExpressions.isEmpty && groupingAttributes.nonEmpty &&
      conf.adaptivePartialAggregationThreshold > 0 &&
      conf.adaptivePartialAggregationThreshold < (1 << conf.fastHashAggregateRowMaxCapacityBit) && {
      child
        .collectUntil(p => p.isInstanceOf[WholeStageCodegenExec] ||
          !p.isInstanceOf[CodegenSupport] ||
          p.isInstanceOf[LeafExecNode]).forall {
        case _: ProjectExec | _: FilterExec | _: ColumnarToRowExec => true
        case _: SerializeFromObjectExec => true
        case _: InputAdapter => true
        // HashAggregateExec, ExpandExec, SortMergeJoinExec ...
        case _ => false
      }
    }
  }
```